Implemented H2DbServer as a factory component
