+++ kura/org.eclipse.kura.core/OSGI-INF/h2dbserver.xml
@@ -12,7 +12,11 @@
       Eurotech
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" immediate="true" name="org.eclipse.kura.core.db.H2DbHelper">
-   <implementation class="org.eclipse.kura.core.db.H2DbHelper"/>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" configuration-policy="require" deactivate="deactivate" immediate="true" modified="updated" name="org.eclipse.kura.core.db.H2DbServer">
+   <implementation class="org.eclipse.kura.core.db.H2DbServer"/>
    <reference bind="setConfigurationService" cardinality="1..1" interface="org.eclipse.kura.configuration.ConfigurationService" name="ConfigurationService" policy="static" unbind="unsetConfigurationService"/>
+   <property name="service.pid" type="String" value="org.eclipse.kura.core.db.H2DbServer"/>
+   <service>
+      <provide interface="org.eclipse.kura.configuration.ConfigurableComponent"/>
+   </service>
 </scr:component>
