+++ kura/org.eclipse.kura.core/OSGI-INF/metatype/org.eclipse.kura.core.db.H2DbServer.xml
@@ -10,7 +10,7 @@
 
 -->
 <MetaData xmlns="http://www.osgi.org/xmlns/metatype/v1.2.0" localization="en_us">
-    <OCD id="org.eclipse.kura.core.db.H2DbHelper" 
+    <OCD id="org.eclipse.kura.core.db.H2DbServer" 
          name="H2DbServer" 
          description="H2 based database service.">
 
@@ -40,10 +40,10 @@
             cardinality="0" 
             required="true"
             default="-tcpPort 9123 -tcpAllowOthers -ifExists"
-            description="Specifies the parameters for the server, see http://www.h2database.com/javadoc/org/h2/tools/Server.html for more details."/>
+            description="Specifies the parameters for the server, see http://www.h2database.com/javadoc/org/h2/tools/Server.html for more details. The listening port must be manually openend in the Firewall configuration section in order to allow external connections."/>
         
         </OCD>
-    <Designate pid="org.eclipse.kura.core.db.H2DbHelper">
-        <Object ocdref="org.eclipse.kura.core.db.H2DbHelper"/>
+    <Designate pid="org.eclipse.kura.core.db.H2DbServer" factoryPid="org.eclipse.kura.core.db.H2DbServer">
+        <Object ocdref="org.eclipse.kura.core.db.H2DbServer"/>
     </Designate>
 </MetaData>
